Themes: Allow Resolver List BG to be fully themed
We are able to trace the activity of the new resolver/chooser through
different
methods, thus leading us to the Java file:
\com\android\internal\app\ChooserActivity.java
Here we see that the exposed "chooser_service_row_background_color" is
available, but not the rest of the activity, so we look into R.layout's,
and we
find chooser_grid to be the only one containing hardcoded
"@color/white" values (as this is framework, we assume this is also
known as
"@android:color/white" to themers).

Expose all "@color/white" values from this file to resolver_list_bg.

Change-Id: I286d92b5d1f672c8adb3c0af1951793521536d90
Signed-off-by: AayushRd7 <aayush35@gmail.com>