OMS-N: Ensure unique type IDs when parsing overlays [12/14]
Always add a block to the idmap for each type in the overlay package
even if there are no matching resources for that type. This ensures that
ResTable::parsePackage will always get a unique ID for each type.

An alternative implementation would be to skip non-matching types when
parsing the overlay package, but that would prevent resolving references
to resources in XML.

Bug: 31052947
Signed-off-by: AayushRd7 <aayush35@gmail.com>

Conflicts:
	libs/androidfw/ResourceTypes.cpp

Change-Id: Iaf34318cc6313c5982e57e4f494aeb05dee5a38d
Signed-off-by: AayushRd7 <aayush35@gmail.com>